More refactoring, trying to keep original naming
