+++ src/main/java/com/splunk/logging/HttpEventCollectorMiddleware.java
@@ -20,11 +20,11 @@ import java.util.List;
 
 /**
  * Splunk http event collector middleware implementation.
- * <p>
+ *
  * A user application can utilize HttpEventCollectorMiddleware to customize the behavior
  * of sending events to Splunk. A user application plugs middleware components to
  * the HttpEventCollectorSender by calling addMiddleware method.
- * <p>
+ *
  * HttpEventCollectorResendMiddleware.java is an example of how middleware can be used.
  */
 
@@ -32,41 +32,20 @@ public class HttpEventCollectorMiddleware implements IHttpEventCollectorMiddlewa
 
     private HttpSenderMiddleware httpSenderMiddleware = null;
 
-    public static class HttpSenderResult {
-        public final HttpSenderSuccess success;
-        public final Exception failed;
-
-        public HttpSenderResult(HttpSenderSuccess success, Exception failed) {
-            this.success = success;
-            this.failed = failed;
-        }
-
-        public HttpSenderResult(HttpSenderSuccess success) {
-            this.success = success;
-            this.failed = null;
-        }
-
-        public HttpSenderResult(Exception failed) {
-            this.success = null;
-            this.failed = failed;
-        }
-    }
-
-    public static class HttpSenderSuccess {
-        public final int statusCode;
-        public final String reply;
-
-        public HttpSenderSuccess(int statusCode, String reply) {
-            this.statusCode = statusCode;
-            this.reply = reply;
-        }
+    /**
+     * An interface that describes an abstract events sender working synchronously/asynchronously.
+     */
+    public interface IHttpSender {
+        void postEvents(final List<HttpEventCollectorEventInfo> events, IHttpSenderCallback callback);
     }
 
     /**
-     * An interface that describes an abstract events sender working synchronously.
+     * Callback methods invoked by events sender.
      */
-    public interface IHttpSender {
-        HttpSenderResult postEvents(final List<HttpEventCollectorEventInfo> events);
+    public interface IHttpSenderCallback {
+        void completed(int statusCode, final String reply);
+
+        void failed(final Exception ex);
     }
 
     /**
@@ -75,36 +54,41 @@ public class HttpEventCollectorMiddleware implements IHttpEventCollectorMiddlewa
     public static abstract class HttpSenderMiddleware implements IHttpSenderMiddleware {
         private HttpSenderMiddleware next;
 
-        public abstract HttpSenderResult postEvents(
-                final List<HttpEventCollectorEventInfo> events,
-                IHttpSender sender);
 
-        protected HttpSenderResult callNext(final List<HttpEventCollectorEventInfo> events,
-                                            IHttpSender sender) {
+        public abstract void postEvents(
+                final List<HttpEventCollectorEventInfo> events,
+                IHttpSender sender,
+                IHttpSenderCallback callback);
+
+        protected void callNext(final List<HttpEventCollectorEventInfo> events,
+                                IHttpSender sender,
+                                IHttpSenderCallback callback) {
             if (next != null) {
-                return next.postEvents(events, sender);
+                next.postEvents(events, sender, callback);
             } else {
-                return sender.postEvents(events);
+                sender.postEvents(events, callback);
             }
         }
     }
 
     /**
      * Post http event collector data
-     *
      * @param events list
+     * @param sender is http sender
+     * @param callback async callback
      */
-    public HttpSenderResult postEvents(final List<HttpEventCollectorEventInfo> events, IHttpSender sender) {
+    public void postEvents(final List<HttpEventCollectorEventInfo> events,
+                           IHttpSender sender,
+                           IHttpSenderCallback callback) {
         if (httpSenderMiddleware == null) {
-            return sender.postEvents(events);
+            sender.postEvents(events, callback);
         } else {
-            return httpSenderMiddleware.postEvents(events, sender);
+            httpSenderMiddleware.postEvents(events, sender, callback);
         }
     }
 
     /**
      * Plug a middleware component to the middleware chain.
-     *
      * @param middleware is a new middleware
      */
     @Override
