+++ src/main/java/com/splunk/logging/HttpEventCollectorMiddlewareSync.java
@@ -20,32 +20,53 @@ import java.util.List;
 
 /**
  * Splunk http event collector middleware implementation.
- *
+ * <p>
  * A user application can utilize HttpEventCollectorMiddleware to customize the behavior
  * of sending events to Splunk. A user application plugs middleware components to
  * the HttpEventCollectorSender by calling addMiddleware method.
- *
+ * <p>
  * HttpEventCollectorResendMiddleware.java is an example of how middleware can be used.
  */
 
-public class HttpEventCollectorMiddlewareAsync implements IHttpEventCollectorMiddleware<HttpEventCollectorMiddlewareAsync.HttpSenderMiddleware> {
+public class HttpEventCollectorMiddlewareSync implements IHttpEventCollectorMiddleware<HttpEventCollectorMiddlewareSync.HttpSenderMiddleware> {
 
     private HttpSenderMiddleware httpSenderMiddleware = null;
 
-    /**
-     * An interface that describes an abstract events sender working asynchronously.
-     */
-    public interface IHttpSender {
-        void postEvents(final List<HttpEventCollectorEventInfo> events, IHttpSenderCallback callback);
+    public static class HttpSenderResult {
+        public final HttpSenderSuccess success;
+        public final Exception failed;
+
+        public HttpSenderResult(HttpSenderSuccess success, Exception failed) {
+            this.success = success;
+            this.failed = failed;
+        }
+
+        public HttpSenderResult(HttpSenderSuccess success) {
+            this.success = success;
+            this.failed = null;
+        }
+
+        public HttpSenderResult(Exception failed) {
+            this.success = null;
+            this.failed = failed;
+        }
+    }
+
+    public static class HttpSenderSuccess {
+        public final int statusCode;
+        public final String reply;
+
+        public HttpSenderSuccess(int statusCode, String reply) {
+            this.statusCode = statusCode;
+            this.reply = reply;
+        }
     }
 
     /**
-     * Callback methods invoked by events sender.
+     * An interface that describes an abstract events sender working synchronously.
      */
-    public interface IHttpSenderCallback {
-        void completed(int statusCode, final String reply);
-
-        void failed(final Exception ex);
+    public interface IHttpSender {
+        HttpSenderResult postEvents(final List<HttpEventCollectorEventInfo> events);
     }
 
     /**
@@ -54,41 +75,36 @@ public class HttpEventCollectorMiddlewareAsync implements IHttpEventCollectorMid
     public static abstract class HttpSenderMiddleware implements IHttpSenderMiddleware {
         private HttpSenderMiddleware next;
 
-
-        public abstract void postEvents(
+        public abstract HttpSenderResult postEvents(
                 final List<HttpEventCollectorEventInfo> events,
-                IHttpSender sender,
-                IHttpSenderCallback callback);
+                IHttpSender sender);
 
-        protected void callNext(final List<HttpEventCollectorEventInfo> events,
-                                IHttpSender sender,
-                                IHttpSenderCallback callback) {
+        protected HttpSenderResult callNext(final List<HttpEventCollectorEventInfo> events,
+                                            IHttpSender sender) {
             if (next != null) {
-                next.postEvents(events, sender, callback);
+                return next.postEvents(events, sender);
             } else {
-                sender.postEvents(events, callback);
+                return sender.postEvents(events);
             }
         }
     }
 
     /**
      * Post http event collector data
+     *
      * @param events list
-     * @param sender is http sender
-     * @param callback async callback
      */
-    public void postEvents(final List<HttpEventCollectorEventInfo> events,
-                           IHttpSender sender,
-                           IHttpSenderCallback callback) {
+    public HttpSenderResult postEvents(final List<HttpEventCollectorEventInfo> events, IHttpSender sender) {
         if (httpSenderMiddleware == null) {
-            sender.postEvents(events, callback);
+            return sender.postEvents(events);
         } else {
-            httpSenderMiddleware.postEvents(events, sender, callback);
+            return httpSenderMiddleware.postEvents(events, sender);
         }
     }
 
     /**
      * Plug a middleware component to the middleware chain.
+     *
      * @param middleware is a new middleware
      */
     @Override
